Simplify Cardano cluster integration test #284
* Rename file to match what we are testing, eg. CardanoCluster and not
LocalCluster
* remove failing assertion that's not relevant anymore: This used to
test we could construct basic Hydra transactions "by hand" but it's
getting more complicated and pretty much useless as it's thoroughly
tested elsewhere
